requirements.txt
flask==1.1.2
requests==2.24.0
wtforms==2.3.1
flask-sqlalchemy==2.4.3
flask-login==0.5.0
cymysql==0.9.14
email-validator==1.1.1
flask-mail==0.9.1
send-email
gunicorn
gevent
